can: isotp: drop single FF transmissions without following CFs
When a SF PDU does not provide enough space the segmentation with FF/CF is
used. The decision is made on the transmission side for all TX_DL values.

On the receiver side we can only check if the sent PDU length requires a
following CF PDU after the initial FF PDU - which is needed in ISO 15765-2.

This patch removes an offset calculation for TX_DL > 8 which is only
required for the transmission(!) side and which was accidentally introduced
on the receiver side on 2013-09-27 11:41:04 in commit b0fe267c
("isotp: update FF_DL length check for SF_DL escape handling")
https://github.com/hartkopp/can-isotp-modules/commit/b0fe267c

This (wrong) offset calculation never caused a real world problem but the
inconsistency showed up in a review for the CAN XL feature implementation.

Signed-off-by: Oliver Hartkopp <socketcan@hartkopp.net>